Delete non-hmac netfilter, move hmac encryption methods directly to netfilter
