+++ SteamKit2/SteamKit2/Networking/Steam3/EnvelopeEncryptedConnection.cs
@@ -10,7 +10,7 @@ namespace SteamKit2
     {
         public EnvelopeEncryptedConnection( IConnection inner, EUniverse universe, ILogContext log, IDebugNetworkListener? debugNetworkListener )
         {
-            this.inner = inner ?? throw new ArgumentNullException( nameof(inner) );
+            this.inner = inner ?? throw new ArgumentNullException( nameof( inner ) );
             this.universe = universe;
             this.log = log ?? throw new ArgumentNullException( nameof( log ) );
             this.debugNetworkListener = debugNetworkListener;
@@ -24,7 +24,9 @@ namespace SteamKit2
         readonly EUniverse universe;
         readonly ILogContext log;
         EncryptionState state;
+#pragma warning disable CA1859 // Use concrete types when possible for improved performance
         INetFilterEncryption? encryption;
+#pragma warning restore CA1859
         IDebugNetworkListener? debugNetworkListener;
 
         public EndPoint? CurrentEndPoint => inner.CurrentEndPoint;
@@ -72,18 +74,18 @@ namespace SteamKit2
 
         void OnNetMsgReceived( object? sender, NetMsgEventArgs e )
         {
-            if (state == EncryptionState.Encrypted)
+            if ( state == EncryptionState.Encrypted )
             {
                 var plaintextData = encryption!.ProcessIncoming( e.Data );
                 NetMsgReceived?.Invoke( this, e.WithData( plaintextData ) );
                 return;
             }
-            
+
             var packetMsg = CMClient.GetPacketMsg( e.Data, log );
 
             if ( packetMsg == null )
             {
-                log.LogDebug( nameof(EnvelopeEncryptedConnection), "Failed to parse message during channel setup, shutting down connection" );
+                log.LogDebug( nameof( EnvelopeEncryptedConnection ), "Failed to parse message during channel setup, shutting down connection" );
                 Disconnect( userInitiated: false );
                 return;
             }
@@ -99,7 +101,7 @@ namespace SteamKit2
 
             if ( !IsExpectedEMsg( packetMsg.MsgType ) )
             {
-                log.LogDebug( nameof(EnvelopeEncryptedConnection), "Rejected EMsg: {0} during channel setup", packetMsg.MsgType );
+                log.LogDebug( nameof( EnvelopeEncryptedConnection ), "Rejected EMsg: {0} during channel setup", packetMsg.MsgType );
                 return;
             }
 
@@ -114,7 +116,7 @@ namespace SteamKit2
                     break;
             }
         }
-        
+
         void HandleEncryptRequest( IPacketMsg packetMsg )
         {
             var request = new Msg<MsgChannelEncryptRequest>( packetMsg );
@@ -122,32 +124,25 @@ namespace SteamKit2
             var connectedUniverse = request.Body.Universe;
             var protoVersion = request.Body.ProtocolVersion;
 
-            log.LogDebug( nameof(EnvelopeEncryptedConnection), "Got encryption request. Universe: {0} Protocol ver: {1}", connectedUniverse, protoVersion );
-            DebugLog.Assert( protoVersion == 1, nameof(EnvelopeEncryptedConnection), "Encryption handshake protocol version mismatch!" );
-            DebugLog.Assert( connectedUniverse == universe, nameof(EnvelopeEncryptedConnection), FormattableString.Invariant( $"Expected universe {universe} but server reported universe {connectedUniverse}" ) );
+            log.LogDebug( nameof( EnvelopeEncryptedConnection ), "Got encryption request. Universe: {0} Protocol ver: {1}", connectedUniverse, protoVersion );
+            DebugLog.Assert( protoVersion == 1, nameof( EnvelopeEncryptedConnection ), "Encryption handshake protocol version mismatch!" );
+            DebugLog.Assert( connectedUniverse == universe, nameof( EnvelopeEncryptedConnection ), FormattableString.Invariant( $"Expected universe {universe} but server reported universe {connectedUniverse}" ) );
+            DebugLog.Assert( request.Payload.Length >= 16, nameof( EnvelopeEncryptedConnection ), "Encryption handshake has a payload that is less than 16 bytes" );
 
-            byte[]? randomChallenge;
-            if ( request.Payload.Length >= 16 )
-            {
-                randomChallenge = request.Payload.ToArray();
-            }
-            else
-            {
-                randomChallenge = null;
-            }
+            var randomChallenge = request.Payload.ToArray();
 
             var publicKey = KeyDictionary.GetPublicKey( connectedUniverse );
 
             if ( publicKey == null )
             {
-                log.LogDebug( nameof(EnvelopeEncryptedConnection), "HandleEncryptRequest got request for invalid universe! Universe: {0} Protocol ver: {1}", connectedUniverse, protoVersion );
+                log.LogDebug( nameof( EnvelopeEncryptedConnection ), "HandleEncryptRequest got request for invalid universe! Universe: {0} Protocol ver: {1}", connectedUniverse, protoVersion );
 
                 Disconnect( userInitiated: false );
                 return;
             }
 
             var response = new Msg<MsgChannelEncryptResponse>();
-            
+
             var tempSessionKey = RandomNumberGenerator.GetBytes( 32 );
             byte[] encryptedHandshakeBlob;
 
@@ -155,20 +150,11 @@ namespace SteamKit2
             {
                 rsa.ImportSubjectPublicKeyInfo( publicKey, out _ );
 
-                var padding = RSAEncryptionPadding.OaepSHA1;
+                var blobToEncrypt = new byte[ tempSessionKey.Length + randomChallenge.Length ];
+                Array.Copy( tempSessionKey, blobToEncrypt, tempSessionKey.Length );
+                Array.Copy( randomChallenge, 0, blobToEncrypt, tempSessionKey.Length, randomChallenge.Length );
 
-                if ( randomChallenge != null )
-                {
-                    var blobToEncrypt = new byte[ tempSessionKey.Length + randomChallenge.Length ];
-                    Array.Copy( tempSessionKey, blobToEncrypt, tempSessionKey.Length );
-                    Array.Copy( randomChallenge, 0, blobToEncrypt, tempSessionKey.Length, randomChallenge.Length );
-
-                    encryptedHandshakeBlob = rsa.Encrypt( blobToEncrypt, padding );
-                }
-                else
-                {
-                    encryptedHandshakeBlob = rsa.Encrypt( tempSessionKey, padding );
-                }
+                encryptedHandshakeBlob = rsa.Encrypt( blobToEncrypt, RSAEncryptionPadding.OaepSHA1 );
             }
 
             var keyCrc = Crc32.Hash( encryptedHandshakeBlob );
@@ -176,15 +162,8 @@ namespace SteamKit2
             response.Write( encryptedHandshakeBlob );
             response.Write( keyCrc );
             response.Write( ( uint )0 );
-            
-            if (randomChallenge != null)
-            {
-                encryption = new NetFilterEncryptionWithHMAC( tempSessionKey, log );
-            }
-            else
-            {
-                encryption = new NetFilterEncryption( tempSessionKey, log );
-            }
+
+            encryption = new NetFilterEncryptionWithHMAC( tempSessionKey, log );
 
             var serialized = response.Serialize();
 
@@ -205,7 +184,7 @@ namespace SteamKit2
         {
             var result = new Msg<MsgChannelEncryptResult>( packetMsg );
 
-            log.LogDebug( nameof(EnvelopeEncryptedConnection), "Encryption result: {0}", result.Body.Result );
+            log.LogDebug( nameof( EnvelopeEncryptedConnection ), "Encryption result: {0}", result.Body.Result );
             DebugLog.Assert( encryption != null, nameof( EnvelopeEncryptedConnection ), "Encryption is null" );
 
             if ( result.Body.Result == EResult.OK && encryption != null )
@@ -215,7 +194,7 @@ namespace SteamKit2
             }
             else
             {
-                log.LogDebug( nameof(EnvelopeEncryptedConnection), "Encryption channel setup failed" );
+                log.LogDebug( nameof( EnvelopeEncryptedConnection ), "Encryption channel setup failed" );
                 Disconnect( userInitiated: false );
             }
         }
